Rework Header
Got alignment issues with previous styling, rework header to be two
containers laid out side by side that each hold columns of information
so that alignment would not be an issue.  Still need to work on right
container for aligning image placeholder correctly, but this is a much
cleaner layout and a simpler solution.