Align top section of page
Align all flex containers on top row so they are spaced
at each end of the container with space between each of the
links.

Align all flex containers in bottom row so they align top to bottom
with each other and the bottom edge of the row

Align all flex containers in bottom left so they are aligned to the
left side of the contaner with appropriate spacing between each.

Make sure all colors match stylesheet given in `./images/02.png`